seperate average-rating to star-rating component
